delimitter is not used in rawBuffermode

+++ forge-reverse-ipc-provider/tsup.config.ts
@@ -3,7 +3,11 @@ import fs from 'node:fs';
 
 export default defineConfig({
 	async onSuccess() {
-		fs.writeFileSync('dist/cjs/package.json', `{"type": "commonjs"}`);
-		fs.writeFileSync('dist/esm/package.json', `{"type": "module"}`);
+		if (fs.existsSync('dist/cjs/')) {
+			fs.writeFileSync('dist/cjs/package.json', `{"type": "commonjs"}`);
+		}
+		if (fs.existsSync('dist/esm/')) {
+			fs.writeFileSync('dist/esm/package.json', `{"type": "module"}`);
+		}
 	},
 });
